Move Graph into it’s own file
