+++ lib/rubrowser/graph.rb
@@ -1,0 +1,11 @@
+require 'tsort'
+
+class Graph < Hash
+  include TSort
+
+  alias tsort_each_node each_key
+
+  def tsort_each_child(node, &block)
+    fetch(node) { [] }.each(&block)
+  end
+end
